feat(telemetry): auto-detach dbg tracer on errs
